packages/client/requirements.txt
socket.io-client>=1.0.0
requests>=2.28.0
psutil>=5.9.0
python-dotenv>=0.19.0
watchdog>=2.1.0